requirements.txt
streamlit==1.29.0
requests==2.31.0
Pillow==10.3.0
pywin32==306; sys_platform == 'win32'
pyproj==3.6.1